refactor: IssueStrategy 파일 분리
